backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pdfplumber==0.11.4
sentence-transformers==3.0.1
faiss-cpu
groq>=0.12.0
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.9.2
